+++ changelog.md
@@ -7,6 +7,7 @@
 * Can pick up cards from table
 * Hand is a type of container (like piles) and behaves similarly to keep things consistent.
 * Cards dragged will always be in front of other elements now
+* When 2+ cards overlap on the board, game will always pick the top one.
 
 ## 0.7